+++ embassy-time/src/duration.rs
@@ -8,47 +8,47 @@ use crate::GCD_1G;
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 /// Represents the difference between two [Instant](struct.Instant.html)s
 pub struct Duration {
-    pub(crate) ticks: u64,
+    pub(crate) ticks: u32,
 }
 
 impl Duration {
     /// The smallest value that can be represented by the `Duration` type.
-    pub const MIN: Duration = Duration { ticks: u64::MIN };
+    pub const MIN: Duration = Duration { ticks: u32::MIN };
     /// The largest value that can be represented by the `Duration` type.
-    pub const MAX: Duration = Duration { ticks: u64::MAX };
+    pub const MAX: Duration = Duration { ticks: u32::MAX };
 
     /// Tick count of the `Duration`.
-    pub const fn as_ticks(&self) -> u64 {
+    pub const fn as_ticks(&self) -> u32 {
         self.ticks
     }
 
     /// Convert the `Duration` to seconds, rounding down.
-    pub const fn as_secs(&self) -> u64 {
+    pub const fn as_secs(&self) -> u32 {
         self.ticks / TICK_HZ
     }
 
     /// Convert the `Duration` to milliseconds, rounding down.
-    pub const fn as_millis(&self) -> u64 {
+    pub const fn as_millis(&self) -> u32 {
         self.ticks * (1000 / GCD_1K) / (TICK_HZ / GCD_1K)
     }
 
     /// Convert the `Duration` to microseconds, rounding down.
-    pub const fn as_micros(&self) -> u64 {
+    pub const fn as_micros(&self) -> u32 {
         self.ticks * (1_000_000 / GCD_1M) / (TICK_HZ / GCD_1M)
     }
 
     /// Creates a duration from the specified number of clock ticks
-    pub const fn from_ticks(ticks: u64) -> Duration {
+    pub const fn from_ticks(ticks: u32) -> Duration {
         Duration { ticks }
     }
 
     /// Creates a duration from the specified number of seconds, rounding up.
-    pub const fn from_secs(secs: u64) -> Duration {
+    pub const fn from_secs(secs: u32) -> Duration {
         Duration { ticks: secs * TICK_HZ }
     }
 
     /// Creates a duration from the specified number of milliseconds, rounding up.
-    pub const fn from_millis(millis: u64) -> Duration {
+    pub const fn from_millis(millis: u32) -> Duration {
         Duration {
             ticks: div_ceil(millis * (TICK_HZ / GCD_1K), 1000 / GCD_1K),
         }
@@ -56,7 +56,7 @@ impl Duration {
 
     /// Creates a duration from the specified number of microseconds, rounding up.
     /// NOTE: Delays this small may be inaccurate.
-    pub const fn from_micros(micros: u64) -> Duration {
+    pub const fn from_micros(micros: u32) -> Duration {
         Duration {
             ticks: div_ceil(micros * (TICK_HZ / GCD_1M), 1_000_000 / GCD_1M),
         }
@@ -64,19 +64,19 @@ impl Duration {
 
     /// Creates a duration from the specified number of nanoseconds, rounding up.
     /// NOTE: Delays this small may be inaccurate.
-    pub const fn from_nanos(micros: u64) -> Duration {
+    pub const fn from_nanos(micros: u32) -> Duration {
         Duration {
             ticks: div_ceil(micros * (TICK_HZ / GCD_1G), 1_000_000_000 / GCD_1G),
         }
     }
 
     /// Creates a duration from the specified number of seconds, rounding down.
-    pub const fn from_secs_floor(secs: u64) -> Duration {
+    pub const fn from_secs_floor(secs: u32) -> Duration {
         Duration { ticks: secs * TICK_HZ }
     }
 
     /// Creates a duration from the specified number of milliseconds, rounding down.
-    pub const fn from_millis_floor(millis: u64) -> Duration {
+    pub const fn from_millis_floor(millis: u32) -> Duration {
         Duration {
             ticks: millis * (TICK_HZ / GCD_1K) / (1000 / GCD_1K),
         }
@@ -84,7 +84,7 @@ impl Duration {
 
     /// Creates a duration from the specified number of microseconds, rounding down.
     /// NOTE: Delays this small may be inaccurate.
-    pub const fn from_micros_floor(micros: u64) -> Duration {
+    pub const fn from_micros_floor(micros: u32) -> Duration {
         Duration {
             ticks: micros * (TICK_HZ / GCD_1M) / (1_000_000 / GCD_1M),
         }
@@ -93,7 +93,7 @@ impl Duration {
     /// Creates a duration corresponding to the specified Hz.
     /// NOTE: Giving this function a hz >= the TICK_HZ of your platform will clamp the Duration to 1
     /// tick. Doing so will not deadlock, but will certainly not produce the desired output.
-    pub const fn from_hz(hz: u64) -> Duration {
+    pub const fn from_hz(hz: u32) -> Duration {
         let ticks = {
             if hz >= TICK_HZ {
                 1
@@ -198,14 +198,14 @@ impl<'a> fmt::Display for Duration {
 }
 
 #[inline]
-const fn div_ceil(num: u64, den: u64) -> u64 {
+const fn div_ceil(num: u32, den: u32) -> u32 {
     (num + den - 1) / den
 }
 
 impl TryFrom<core::time::Duration> for Duration {
-    type Error = <u64 as TryFrom<u128>>::Error;
+    type Error = <u32 as TryFrom<u128>>::Error;
 
-    /// Converts using [`Duration::from_micros`]. Fails if value can not be represented as u64.
+    /// Converts using [`Duration::from_micros`]. Fails if value can not be represented as u32.
     fn try_from(value: core::time::Duration) -> Result<Self, Self::Error> {
         Ok(Self::from_micros(value.as_micros().try_into()?))
     }
@@ -214,6 +214,6 @@ impl TryFrom<core::time::Duration> for Duration {
 impl From<Duration> for core::time::Duration {
     /// Converts using [`Duration::as_micros`].
     fn from(value: Duration) -> Self {
-        core::time::Duration::from_micros(value.as_micros())
+        core::time::Duration::from_micros(value.as_micros() as u64)
     }
 }
